+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 - "payment" account type for Pix related resources
 - missing parameters to Boleto, BrcodePayment, DictKey, Event, Invoice, Transfer and Workspace resources
 - Workspace.update() to allow parameter updates
+- Invoice.Payment sub-resource to allow retrieval of invoice payment information
 
 ## [2.4.1] - 2021-03-19
 ### Fixed